extracted common shared aiopg_utils
